# [JSM] Fix periodic failures of Selenium test 'selenium_agent_add_comment'